pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "email-classifier"
version = "0.1.0"
description = "FastAPI app that classifies emails/docs (produtivo x improdutivo) using Google Gemini."
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
  "fastapi[standard]>=0.116,<1.0", 
  "pydantic-settings>=2.10,<3.0",
  "pypdf>=4.2",
  "google-genai>=1.30",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3",
  "pytest-cov>=5.0",
  "ruff>=0.6",
  "black>=24.8",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
pythonpath = ["."]

[tool.ruff]
line-length = 100
target-version = "py313"
lint.select = ["E", "F", "I", "UP"] 

[tool.black]
line-length = 100
target-version = ["py313"]
